requirements.txt
streamlit==1.41.0
requests==2.32.0
python-dotenv==1.0.0
azure-identity==1.18.0
pydantic==2.6.0
pyarrow==15.0.0
pandas==2.2.0
aiohttp==3.10.0
pyyaml==6.0.1
